constexpr

+++ arangod/Indexes/Index.cpp
@@ -205,7 +205,7 @@ bool Index::validateHandle(char const* key, size_t* split) {
     return false;
   }
 
-  if (p - key > TRI_COL_NAME_LENGTH) {
+  if (static_cast<size_t>(p - key) > TRI_COL_NAME_LENGTH) {
     return false;
   }
